feature: add all agent-specific config blocks to nomad-types (public data only)
