+++ agents/processor/Cargo.toml
@@ -21,6 +21,7 @@ tracing-subscriber = "0.2.15"
 rocksdb = { git = "https://github.com/rust-rocksdb/rust-rocksdb" }
 nomad-xyz-configuration = { git = "https://github.com/nomad-xyz/configuration", rev = "0410551" }
 
+nomad-types = { path = "../../nomad-types" }
 nomad-core = { path = "../../nomad-core" }
 nomad-base = { path = "../../nomad-base" }
 paste = "1.0.5"
