+++ nomad-types/src/agent/relayer.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RelayerConfig {
+    pub interval: u64,
+}
