+++ nomad-types/src/agent/updater.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UpdaterSettingsBlock {
+    pub interval: u64,
+}
